Bugfix: Recalculate bootloader hash when flash-size is changed.
Recompute and update the bootloader SHA256 hash when the header is
modified, eg. when the flash_size has been changed.

This avoids the warnings about incorrect sha hash when booting the
esp32 device after changing the flash_size with the "--flash-size"
option.

Also, update the image_file module to use the ImageFormat class to
read and manipulate the bootloader images.