esptool_io.py: Add ESPToolProgressBar class for monitoring esptool.py progress.
Refactor and simplification of progressbar code. The ESPToolProgressBar
class now monitors the input stream for progress messages and updates a
progress bar accordingly. The class can be used as a context manager to
run the progress bar in a separate thread while executing the body of
the `with` block. The `run()` method can be used to run the
progress bar in the current thread.