firmware_fileio: Add Partition.truncate() method.
Partition.truncate() will erase the remainder of a partition if acting on an
ESP32 device flash storage. If acting on a firmware file, it will truncate the
file to the new size if this is the last partition in the firmware file (eg. the
'factory' partition).

After writing new data to a partition, call the truncate() method to ensure the
rest of the partition is erased.