argtypes: Make ArgList and PartList concrete classes.
Change `ArgList` and `PartList` from type declarations to classes
with constructors that parse the input string into a
list of lists or list of tuples, replacing the `arglist()` and
`partlist()` functions.

Also change the `numeric_arg()` function to a class `IntArg` that
inherits from `int` and parses the input string into an integer value
with an optional trailing unit suffix.