Add --fs option to manipulate files on littlefs filesystems.
This commit adds a new option to the mp_image_tool_esp32 tool to
manipulate files on littlefs filesystems. The new option is called
`--fs` and can be used to list, copy, move, delete, and create files
and directories on the filesystem images in the firmware image or
on the device.

Uses littlefs-python to interact with the filesystem images.