Make littlefs an optional dependency.
This change makes the `littlefs-python` package an optional dependency
for the `fs` command in the `mp-image-tool-esp32` package. This is
achieved by moving the `littlefs-python` package from the `dependencies`
section to the `project.optional-dependencies` section in the
`pyproject.toml` file. The `littlefs` extra is now required to install
the `littlefs-python` package.

pyproject.toml:
- move the `littlefs-python` package to the
  `project.optional-dependencies` section.
- use dependency_groups for the test and typing environments.
- Remove dependencies from the tox configuration and use the
  test dependency_group instead.